Tests/LibWeb/Text/expected/indexeddb-upgrade-close-event-ordering.txt

setup.upgradeneeded
setup.success
connA.versionchange
connA.close() called
upgrade.upgradeneeded
upgrade.success
DONE
